c file
This C program implements the Tower of Hanoi problem using stacks. The Tower of Hanoi is a mathematical puzzle where we have three rods and n disks. The objective is to move the entire stack of disks from one rod to another.
The program uses stacks to simulate the rods and performs the Tower of Hanoi operation iteratively. It prompts the user to input the number of disks, then prints each move required to solve the Tower of Hanoi puzzle.